deviceManager
