+++ VarkaAPP/Additional Files /Extensions/CGPath+Extension.swift
@@ -15,26 +15,15 @@ extension CGPath {
         
         if screenIsSquare {
             
-            
-            
-
             path.move(to: CGPoint(x: 0, y: 0))
-
-            
             path.addLine(to: CGPoint(x: (frame.width / 2) - 50, y: 0))
-            
-
             path.addQuadCurve(to: CGPoint(x: (frame.width / 2) - 40, y: 10),
                               controlPoint: CGPoint(x: (frame.width / 2) - 40, y: 0))
-            
             path.addArc(withCenter: CGPoint(x: frame.width / 2, y: 0),
                         radius: 41, startAngle: CGFloat(3.06),
                         endAngle: CGFloat(0.224), clockwise: false)
-
             path.addQuadCurve(to: CGPoint(x: (frame.width / 2) + 50, y: 0),
                               controlPoint: CGPoint(x: (frame.width / 2) + 40, y: 0))
-            
-            
             path.addLine(to: CGPoint(x: frame.width, y: 0))
             path.addLine(to: CGPoint(x: frame.width, y: frame.height))
             path.addLine(to: CGPoint(x: 0, y: frame.height))
@@ -57,7 +46,7 @@ extension CGPath {
             path.addArc(withCenter: CGPoint(x: frame.width / 2, y: 10),
                         radius: 44, startAngle: CGFloat(Double.pi),
                         endAngle: CGFloat(0), clockwise: false)
-            path.addQuadCurve(to: CGPoint(x: (frame.width / 2) + 54, y: 0),
+            path.addQuadCurve(to: CGPoint(x: (frame.width / 2) + 55, y: 0),
                               controlPoint: CGPoint(x: (frame.width / 2) + 44, y: 0))
             path.addLine(to: CGPoint(x: frame.width - sideIndent - 10, y: 0))
             path.addQuadCurve(to: CGPoint(x: frame.width - sideIndent, y: 10),
