+++ VarkaAPP/Additional Files /Helpers/Models.swift
@@ -1,0 +1,40 @@
+//
+//  Models.swift
+//  VarkaAPP
+//
+//  Created by Никита Гвоздиков on 04.04.2021.
+//
+
+import Foundation
+
+enum Models : String {
+    
+    case simulator          = "simulator/sandbox",
+         iPhone4            = "iPhone 4",
+         iPhone4S           = "iPhone 4S",
+         iPhone5            = "iPhone 5",
+         iPhone5S           = "iPhone 5S",
+         iPhone5C           = "iPhone 5C",
+         iPhone6            = "iPhone 6",
+         iPhone6Plus        = "iPhone 6 Plus",
+         iPhone6S           = "iPhone 6S",
+         iPhone6SPlus       = "iPhone 6S Plus",
+         iPhoneSE           = "iPhone SE",
+         iPhone7            = "iPhone 7",
+         iPhone7Plus        = "iPhone 7 Plus",
+         iPhone8            = "iPhone 8",
+         iPhone8Plus        = "iPhone 8 Plus",
+         iPhoneX            = "iPhone X",
+         iPhoneXS           = "iPhone XS",
+         iPhoneXSMax        = "iPhone XS Max",
+         iPhoneXR           = "iPhone XR",
+         iPhone11           = "iPhone 11",
+         iPhone11Pro        = "iPhone 11 Pro",
+         iPhone11ProMax     = "iPhone 11 Pro Max",
+         iPhoneSE2          = "iPhone SE 2nd gen",
+         iPhone12Mini       = "iPhone 12 mini",
+         iPhone12           = "iPhone 12",
+         iPhone12Pro        = "iPhone 12 Pro",
+         iPhone12ProMax     = "iPhone 12 Pro Max",
+         unrecognized       = "?unrecognized?"
+}
